requirements.txt
--index-url https://download.pytorch.org/whl/cu117

torch=2.0.1+cu117
torchvision=0.15.2+cu117
pandas==2.0.3
matplotlib==3.7.2
psutil
scikit-learn==1.3.0
pyyaml==6.0.1
pillow==10.0.0
tqdm
timm
umap-learn